Main/requirements.txt
streamlit
pandas
PyPDF2
scikit-learn
selenium
webdriver-manager
beautifulsoup4
google-generativeai
python-dotenv
psycopg2-binary